# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Volcanion Stress Test Tool API
  description: |
    A powerful HTTP load testing and stress testing tool built in Go.
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **JWT Bearer Token**: Obtain a token via `/api/v1/auth/login` and include it in the `Authorization` header.
    2. **API Key**: Include your API key in the `X-API-Key` header.
    
    ## Rate Limiting
    
    API endpoints are rate-limited to 100 requests per minute per IP address.
    When rate limited, you'll receive a `429 Too Many Requests` response.
    
    ## Versioning
    
    All API endpoints are versioned under `/api/v1/`. The current version is returned 
    in the `X-API-Version` response header.
    
  version: 1.0.0
  contact:
    name: Volcanion Team
    url: https://github.com/volcanion-company/volcanion-stress-test-tool
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.volcanion.io
    description: Production server (example)
...
paths:
...
  /api/v1/test-plans:
    post:
      summary: Create Test Plan
      description: Create a new test plan configuration
      operationId: createTestPlan
      tags:
        - Test Plans
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTestPlanRequest'
      responses:
        '201':
          description: Test plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      summary: List Test Plans
      description: Get all test plans
      operationId: listTestPlans
      tags:
        - Test Plans
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of test plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TestPlan'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/test-plans/{id}:
    get:
      summary: Get Test Plan
      description: Get a specific test plan by ID
      operationId: getTestPlan
      tags:
        - Test Plans
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test plan details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/test-runs/{id}:
    get:
      summary: Get Test Run
      description: Get a specific test run by ID
      operationId: getTestRun
      tags:
        - Test Runs
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Test run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestRun'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
...
    CreateTestPlanRequest:
      type: object
      required:
        - name
        - target_url
        - method
        - users
        - duration_sec
      properties:
        name:
          type: string
          description: Test plan name
          example: API Load Test
        target_url:
          type: string
          format: uri
          description: Target URL to test
          example: https://api.example.com/endpoint
        method:
          type: string
          enum: [GET, POST, PUT, PATCH, DELETE]
          description: HTTP method
        headers:
          type: object
          additionalProperties:
            type: string
          description: Custom HTTP headers
        body:
          type: string
          description: Request body (for POST/PUT/PATCH)
        users:
          type: integer
          minimum: 1
          description: Number of concurrent virtual users
          example: 100
        ramp_up_sec:
          type: integer
          minimum: 0
          description: Ramp-up period in seconds
          default: 0
        duration_sec:
          type: integer
          minimum: 1
          description: Test duration in seconds
          example: 60
        timeout_ms:
          type: integer
          minimum: 0
          description: Request timeout in milliseconds
          default: 30000
        target_rps:
          type: integer
          minimum: 0
          description: Target requests per second (0 = unlimited)
          default: 0
        rate_pattern:
          type: string
          enum: [fixed, step, spike, ramp]
          default: fixed
        sla:
          $ref: '#/components/schemas/SLAConfig'
    TestPlan:
      allOf:
        - $ref: '#/components/schemas/CreateTestPlanRequest'
        - type: object
          properties:
            id:
              type: string
            created_at:
              type: string
              format: date-time
    TestRun:
      type: object
      properties:
        id:
          type: string
        plan_id:
          type: string
        status:
          type: string
          enum: [running, completed, cancelled, failed]
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        metrics:
          $ref: '#/components/schemas/Metrics'
    Metrics:
      type: object
      properties:
        run_id:
          type: string
        total_requests:
          type: integer
        successful_requests:
          type: integer
        failed_requests:
          type: integer
        avg_latency_ms:
          type: number
          format: double
        min_latency_ms:
          type: number
          format: double
        max_latency_ms:
          type: number
          format: double
        p50_latency_ms:
          type: number
          format: double
        p90_latency_ms:
          type: number
          format: double
        p95_latency_ms:
          type: number
          format: double
        p99_latency_ms:
          type: number
          format: double
        requests_per_second:
          type: number
          format: double
        error_rate:
          type: number
          format: double
        status_codes:
          type: object
          additionalProperties:
            type: integer
...
    SLAConfig:
      type: object
      properties:
        max_p95_latency:
          type: number
          description: Maximum P95 latency in milliseconds
        max_error_rate:
          type: number
          description: Maximum error rate (0.0 - 1.0)
        min_rps:
          type: number
          description: Minimum requests per second
...